custom hook useWindowSize, props to pagination added
